Moved mixin as MultipleLinesComposition.
